third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_write_file=tests1-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <?
  assert_equals: expected "#document\\n| <!-- ? -->\\n| <html>\\n|   <head>\\n|   <body>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>"
[FAIL] <?COMMENT?>
  assert_equals: expected "#document\\n| <!-- ?COMMENT? -->\\n| <html>\\n|   <head>\\n|   <body>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>"
[FAIL] <?COM--MENT?>
  assert_equals: expected "#document\\n| <!-- ?COM--MENT? -->\\n| <html>\\n|   <head>\\n|   <body>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>"
Harness: the test ran to completion.
